rand: make pseudo-random variants a bit more variable
Until now all pseudo-random variants were always strings. Since a true
variant can contain any _complete_ type, let's use at least any
_basic_ type, for now. Extending this to any complete type must wait
until we can generate arbitrary signatures.
